♻️ Dedup mint-handshake-then-dispatch across enqueue/promote paths
enqueueAndDispatchJob, promoteAndDispatchJob, and promoteScheduledJobs
each independently declared HANDSHAKE_EXPIRY_SECONDS and repeated the
same createHandshakeToken-then-dispatchJob sequence. Extracted
dispatchJobWithHandshake into dispatchJob.server.ts (the existing home
for the dispatch primitive) and had all three call it.

Also folded promoteScheduledJobs' dispatchPromotedJob into the shared
helper directly — it had no real callers anywhere in the repo or the
extension modules (verified via search), despite being wildcard
re-exported from the package's public barrel.
