+++ packages/scms-server/src/backend/jobs/enqueue/dispatchJob.server.ts
@@ -1,5 +1,9 @@
+import { getConfig } from '../../../app-config.server.js';
+import { createHandshakeToken } from '../../sign.handshake.server.js';
 import { sendJobMessage } from './pgmq/jobQueue.server.js';
-import type { JobQueueMessage } from './pgmq/types.js';
+import type { JobQueueMessage, JobQueueSendResult } from './pgmq/types.js';
+
+const HANDSHAKE_EXPIRY_SECONDS = 4 * 60 * 60;
 
 /**
  * Enqueue a job message onto the pgmq job queue.
@@ -11,4 +15,25 @@ import type { JobQueueMessage } from './pgmq/types.js';
  */
 export async function dispatchJob(message: JobQueueMessage) {
   return sendJobMessage(message, { idempotencyKey: message.job_id });
+}
+
+/**
+ * Mint a fresh handshake token for a job and dispatch it. Shared by every
+ * enqueue/promote path (enqueueAndDispatchJob, promoteAndDispatchJob,
+ * promoteScheduledJobs) — they all need the same mint-then-dispatch sequence
+ * with the same expiry.
+ */
+export async function dispatchJobWithHandshake(job: {
+  id: string;
+  job_type: string;
+}): Promise<JobQueueSendResult> {
+  const config = await getConfig();
+  const handshake = createHandshakeToken(
+    job.id,
+    job.job_type,
+    config.api.handshakeIssuer,
+    config.api.handshakeSigningSecret,
+    Math.floor(Date.now() / 1000) + HANDSHAKE_EXPIRY_SECONDS,
+  );
+  return dispatchJob({ job_id: job.id, job_type: job.job_type, handshake });
 }
