+++ packages/scms-server/src/backend/jobs/enqueue/promoteAndDispatchJob.server.spec.ts
@@ -2,7 +2,7 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { JobStatus } from '@curvenote/scms-db';
 
-const mockDispatchJob = vi.fn();
+const mockDispatchJobWithHandshake = vi.fn();
 const mockFindUnique = vi.fn();
 const mockUpdate = vi.fn();
 const mockUpdateMany = vi.fn();
@@ -11,22 +11,12 @@ const mockPrisma = {
   job: { findUnique: mockFindUnique, update: mockUpdate, updateMany: mockUpdateMany },
 };
 
-vi.mock('../../../app-config.server.js', () => ({
-  getConfig: vi.fn(async () => ({
-    api: { handshakeIssuer: 'issuer', handshakeSigningSecret: 'secret' },
-  })),
-}));
-
-vi.mock('../../sign.handshake.server.js', () => ({
-  createHandshakeToken: vi.fn(() => 'handshake-token'),
-}));
-
 vi.mock('../../prisma.server.js', () => ({
   getPrismaClient: vi.fn(async () => mockPrisma),
 }));
 
 vi.mock('./dispatchJob.server.js', () => ({
-  dispatchJob: (...args: unknown[]) => mockDispatchJob(...args),
+  dispatchJobWithHandshake: (...args: unknown[]) => mockDispatchJobWithHandshake(...args),
 }));
 
 const { promoteAndDispatchJob } = await import('./promoteAndDispatchJob.server.js');
@@ -37,7 +27,7 @@ describe('promoteAndDispatchJob', () => {
     mockFindUnique.mockResolvedValue({ id: 'dep-1', job_type: 'CHECK', status: JobStatus.BLOCKED });
     mockUpdate.mockResolvedValue({});
     mockUpdateMany.mockResolvedValue({ count: 1 });
-    mockDispatchJob.mockResolvedValue({ messageId: '1' });
+    mockDispatchJobWithHandshake.mockResolvedValue({ messageId: '1' });
   });
 
   it('promotes BLOCKED to QUEUED and dispatches on success', async () => {
@@ -47,12 +37,12 @@ describe('promoteAndDispatchJob', () => {
       where: { id: 'dep-1' },
       data: { status: JobStatus.QUEUED },
     });
-    expect(mockDispatchJob).toHaveBeenCalledTimes(1);
+    expect(mockDispatchJobWithHandshake).toHaveBeenCalledTimes(1);
     expect(mockUpdateMany).not.toHaveBeenCalled();
   });
 
   it('reverts to BLOCKED when dispatch fails, without throwing', async () => {
-    mockDispatchJob.mockRejectedValueOnce(new Error('pgmq unavailable'));
+    mockDispatchJobWithHandshake.mockRejectedValueOnce(new Error('pgmq unavailable'));
 
     await expect(promoteAndDispatchJob('dep-1')).resolves.toBeUndefined();
 
@@ -68,6 +58,6 @@ describe('promoteAndDispatchJob', () => {
     await promoteAndDispatchJob('dep-1');
 
     expect(mockUpdate).not.toHaveBeenCalled();
-    expect(mockDispatchJob).not.toHaveBeenCalled();
+    expect(mockDispatchJobWithHandshake).not.toHaveBeenCalled();
   });
 });
