+++ packages/scms-server/src/backend/jobs/enqueue/promoteScheduledJobs.server.spec.ts
@@ -2,7 +2,7 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { JobStatus } from '@curvenote/scms-db';
 
-const mockDispatchJob = vi.fn();
+const mockDispatchJobWithHandshake = vi.fn();
 const mockUpdateMany = vi.fn();
 const mockQueryRaw = vi.fn();
 
@@ -15,22 +15,12 @@ const mockPrisma = {
   job: { updateMany: mockUpdateMany },
 };
 
-vi.mock('../../../app-config.server.js', () => ({
-  getConfig: vi.fn(async () => ({
-    api: { handshakeIssuer: 'issuer', handshakeSigningSecret: 'secret' },
-  })),
-}));
-
-vi.mock('../../sign.handshake.server.js', () => ({
-  createHandshakeToken: vi.fn(() => 'handshake-token'),
-}));
-
 vi.mock('../../prisma.server.js', () => ({
   getPrismaClient: vi.fn(async () => mockPrisma),
 }));
 
 vi.mock('./dispatchJob.server.js', () => ({
-  dispatchJob: (...args: unknown[]) => mockDispatchJob(...args),
+  dispatchJobWithHandshake: (...args: unknown[]) => mockDispatchJobWithHandshake(...args),
 }));
 
 const { promoteScheduledJobs } = await import('./promoteScheduledJobs.server.js');
@@ -42,7 +32,7 @@ describe('promoteScheduledJobs', () => {
       { id: 'job-1', job_type: 'LOOPBACK' },
       { id: 'job-2', job_type: 'LOOPBACK' },
     ]);
-    mockDispatchJob.mockResolvedValue({ messageId: '1' });
+    mockDispatchJobWithHandshake.mockResolvedValue({ messageId: '1' });
     mockUpdateMany.mockResolvedValue({ count: 1 });
   });
 
@@ -50,19 +40,19 @@ describe('promoteScheduledJobs', () => {
     const result = await promoteScheduledJobs(2);
 
     expect(result).toEqual({ claimed: 2, dispatched: 2, dispatchFailed: 0 });
-    expect(mockDispatchJob).toHaveBeenCalledTimes(2);
+    expect(mockDispatchJobWithHandshake).toHaveBeenCalledTimes(2);
     expect(mockUpdateMany).not.toHaveBeenCalled();
   });
 
   it('reverts to SCHEDULED and continues when dispatch fails for one job', async () => {
-    mockDispatchJob
+    mockDispatchJobWithHandshake
       .mockRejectedValueOnce(new Error('pgmq unavailable'))
       .mockResolvedValueOnce({ messageId: '2' });
 
     const result = await promoteScheduledJobs(2);
 
     expect(result).toEqual({ claimed: 2, dispatched: 1, dispatchFailed: 1 });
-    expect(mockDispatchJob).toHaveBeenCalledTimes(2);
+    expect(mockDispatchJobWithHandshake).toHaveBeenCalledTimes(2);
     expect(mockUpdateMany).toHaveBeenCalledWith({
       where: { id: 'job-1', status: JobStatus.QUEUED },
       data: expect.objectContaining({ status: JobStatus.SCHEDULED }),
