+++ packages/scms-server/src/backend/jobs/enqueue/promoteScheduledJobs.server.ts
@@ -1,27 +1,8 @@
 import { JobStatus, Prisma } from '@curvenote/scms-db';
-import { getConfig } from '../../../app-config.server.js';
 import { getPrismaClient } from '../../prisma.server.js';
-import { createHandshakeToken } from '../../sign.handshake.server.js';
-import { dispatchJob } from './dispatchJob.server.js';
+import { dispatchJobWithHandshake } from './dispatchJob.server.js';
 
-const HANDSHAKE_EXPIRY_SECONDS = 4 * 60 * 60;
 const DEFAULT_PROMOTE_LIMIT = 100;
-
-async function dispatchPromotedJob(job: { id: string; job_type: string }): Promise<void> {
-  const config = await getConfig();
-  const handshake = createHandshakeToken(
-    job.id,
-    job.job_type,
-    config.api.handshakeIssuer,
-    config.api.handshakeSigningSecret,
-    Math.floor(Date.now() / 1000) + HANDSHAKE_EXPIRY_SECONDS,
-  );
-  await dispatchJob({
-    job_id: job.id,
-    job_type: job.job_type,
-    handshake,
-  });
-}
 
 async function revertPromotedJobToScheduled(jobId: string, nowIso: string): Promise<void> {
   const prisma = await getPrismaClient();
@@ -68,7 +49,7 @@ export async function promoteScheduledJobs(limit = DEFAULT_PROMOTE_LIMIT): Promi
 
   for (const row of promoted) {
     try {
-      await dispatchPromotedJob(row);
+      await dispatchJobWithHandshake(row);
       dispatched += 1;
     } catch (err) {
       dispatchFailed += 1;
@@ -89,5 +70,3 @@ export async function promoteScheduledJobs(limit = DEFAULT_PROMOTE_LIMIT): Promi
   });
   return { claimed: promoted.length, dispatched, dispatchFailed };
 }
-
-export { dispatchPromotedJob };
